# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Sensity Parking APIs
  description: Sensity Parking APIs 
# during dev, should point to your local machine
host: 0.0.0.0:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
...
  /customers/{orgid}/sites/{siteid}/parkingpolicies/{parkingpolicyid}:
    x-swagger-router-controller: parking_policy
    # Get a parking policy 
    get:
      tags:
        - ParkingPolicy
      summary: Get a parking policy in a site
      description: Get a specific parking policy in a site
      operationId: getParkingPolicy
      parameters:
        - name: orgid
          in: path
          description: The org ID
          required: true
          type: string
        - name: siteid
          in: path
          description: The site ID
          required: true
          type: string
        - name: parkingpolicyid
          in: path
          description: The parking policy ID
          required: true
          type: string        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetParkingPolicyResponse"
        "401":
          description: Unauthorized to access report
          schema:
            $ref: "#/definitions/ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "405":
          description: Method Not Allowed
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "502":
          description: Bad Gateway
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/ErrorResponse"
        "504":
          description: Gateway Timeout
          schema:
            $ref: "#/definitions/ErrorResponse"
        "408":
          description: Request Timedout
          schema:
            $ref: "#/definitions/ErrorResponse"

    # Delete a policy category
    delete:
      tags:
        - ParkingPolicy
      summary: Delete a parking policy in a site
      description: Delete a parking policy in a site
      operationId: deleteParkingPolicy
      parameters:
        - name: orgid
          in: path
          description: The customer ID
          required: true
          type: string
        - name: siteid
          in: path
          description: The site ID
          required: true
          type: string
        - name: parkingpolicyid
          in: path
          description: The parking policy ID
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteParkingPolicyResponse"
        "401":
          description: Unauthorized to access report
          schema:
            $ref: "#/definitions/ErrorResponse" 
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse" 
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse" 
        "405":
          description: Method Not Allowed
          schema:
            $ref: "#/definitions/ErrorResponse" 
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse" 
        "502":
          description: Bad Gateway
          schema:
            $ref: "#/definitions/ErrorResponse" 
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/ErrorResponse" 
        "504":
          description: Gateway Timeout
          schema:
            $ref: "#/definitions/ErrorResponse" 
        "408":
          description: Request Timedout
          schema:
            $ref: "#/definitions/ErrorResponse" 

    # Update a parking policy
    put:
      tags:
        - ParkingPolicy
      summary: Update a parking policy
      description: Update a parking policy
      operationId: updateParkingPolicy
      parameters:
        - name: orgid
          in: path
          description: The customer ID
          required: true
          type: string
        - name: siteid
          in: path
          description: The site ID
          required: true
          type: string
        - name: parkingpolicyid
          in: path
          description: The parking policy ID
          required: true
          type: string
        - name: policy
          in: body
          description: Updated a parking policy
          required: true
          schema:
            $ref: "#/definitions/NewParkingPolicy"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UpdateParkingPolicyResponse"
        "401":
          description: Unauthorized to access report
          schema:
            $ref: "#/definitions/ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "405":
          description: Method Not Allowed
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "502":
          description: Bad Gateway
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/ErrorResponse"
        "504":
          description: Gateway Timeout
          schema:
            $ref: "#/definitions/ErrorResponse"
        "408":
          description: Request Timedout
          schema:
            $ref: "#/definitions/ErrorResponse" 

   # Create, get all parking policy in a site
  /customers/{orgid}/sites/{siteid}/parkingpolicies:
    x-swagger-router-controller: parking_policy

    # Create a parking policy
    post:
      tags:
        - ParkingPolicy
      summary: Create a parking policy in a site
      description: Create a parking policy in a site
      operationId: createParkingPolicy
      parameters:
        - name: orgid
          in: path
          description: The org ID
          required: true
          type: string
        - name: siteid
          in: path
          description: The site ID
          required: true
          type: string
        - name: policy
          in: body
          description: New Parking Policy Details          
          schema:
            $ref: "#/definitions/NewParkingPolicy"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CreateParkingPolicyResponse"
        "401":
          description: Unauthorized to access report
          schema:
            $ref: "#/definitions/ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "405":
          description: Method Not Allowed
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "502":
          description: Bad Gateway
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/ErrorResponse"
        "504":
          description: Gateway Timeout
          schema:
            $ref: "#/definitions/ErrorResponse"
        "408":
          description: Request Timedout
          schema:
            $ref: "#/definitions/ErrorResponse"  

    # Get all parking policies
    get:
      tags:
        - ParkingPolicy
      summary: Get all parking policies in a site
      description: Get all parking policies in a site
      operationId: getAllParkingPolicy
      parameters:
        - name: orgid
          in: path
          description: The org ID
          required: true
          type: string
        - name: siteid
          in: path
          description: The site ID
          required: true
          type: string              
      responses:
        "200":
          description: Success
          schema:
            type: array
            items: 
              # a pointer to a definition
              $ref: "#/definitions/getAllParkingPolicyResponse"
        "401":
          description: Unauthorized to access report
          schema:
            $ref: "#/definitions/ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "405":
          description: Method Not Allowed
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "502":
          description: Bad Gateway
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/ErrorResponse"
        "504":
          description: Gateway Timeout
          schema:
            $ref: "#/definitions/ErrorResponse"
        "408":
          description: Request Timedout
          schema:
            $ref: "#/definitions/ErrorResponse"  

    # Group-Policy Assocation in a site
...
  /customers/{orgid}/sites/{siteid}/parkingpolicies/{parkingpolicyid}/version/{version}:
    x-swagger-router-controller: parking_policy

    # Get a group-policy based on version
    get:
      tags:
        - ParkingPolicy
      summary: Get a parking policy in a site
      description: Get a specific parking policy in a site
      operationId: getParkingPolicyVersion
      parameters:
        - name: orgid
          in: path
          description: The org ID
          required: true
          type: string
        - name: siteid
          in: path
          description: The site ID
          required: true
          type: string
        - name: parkingpolicyid
          in: path
          description: The parking policy ID
          required: true
          type: string
        - name: version
          in: path
          description: The version Number
          required: true
          type: integer    
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetParkingPolicyResponseByVersion"
        "401":
          description: Unauthorized to access report
          schema:
            $ref: "#/definitions/ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "405":
          description: Method Not Allowed
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "502":
          description: Bad Gateway
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/ErrorResponse"
        "504":
          description: Gateway Timeout
          schema:
            $ref: "#/definitions/ErrorResponse"
        "408":
          description: Request Timedout
          schema:
            $ref: "#/definitions/ErrorResponse" 

     # Group-Policy based on versionhistory of a site
  /customers/{orgid}/sites/{siteid}/parkingpolicies/{parkingpolicyid}/versionhistory:
    x-swagger-router-controller: parking_policy

    # Get a group-policy based on versionhistory
    get:
      tags:
        - ParkingPolicy
      summary: Get a parking policy in a site
      description: Get a specific parking policy in a site
      operationId: getAllVersionsOfParkingPolicy
      parameters:
        - name: orgid
          in: path
          description: The org ID
          required: true
          type: string
        - name: siteid
          in: path
          description: The site ID
          required: true
          type: string
        - name: parkingpolicyid
          in: path
          description: The parking policy ID
          required: true
          type: string   
      responses:
        "200":
          description: Success
          schema:
            type: array
            items: 
              # a pointer to a definition
              $ref: "#/definitions/GetParkingPolicyResponseByVersionHistory"
        "401":
          description: Unauthorized to access report
          schema:
            $ref: "#/definitions/ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "405":
          description: Method Not Allowed
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "502":
          description: Bad Gateway
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/ErrorResponse"
        "504":
          description: Gateway Timeout
          schema:
            $ref: "#/definitions/ErrorResponse"
        "408":
          description: Request Timedout
          schema:
            $ref: "#/definitions/ErrorResponse"  

      # Group-Policy based on from and to timestamps of a site
  /customers/{orgid}/sites/{siteid}/parkinggroups/{parkinggroupid}/activepolicies/fromtime/{fromtime}/totime/{totime}:
    x-swagger-router-controller: parking_policy

    # Get a group-policy based on timestamps
    get:
      tags:
        - ParkingPolicy
      summary: Get a parking policy of a site within timeline
      description: Get a specific parking policy of a site within timeline
      operationId: getAllActiveParkingPolicyForPeriod
      parameters:
        - name: orgid
          in: path
          description: The org ID
          required: true
          type: string
        - name: siteid
          in: path
          description: The site ID
          required: true
          type: string
        - name: parkinggroupid
          in: path
          description: The Parking Group ID
          required: true
          type: string
        - name: fromtime
          in: path
          description: The From Time
          required: true
          type: string 
        - name: totime
          in: path
          description: The To Time
          required: true
          type: string    
      responses:
        "200":
          description: Success
          schema:
            type: array
            items: 
              # a pointer to a definition
              $ref: "#/definitions/GetParkingPolicyResponseByTimeline"
        "401":
          description: Unauthorized to access report
          schema:
            $ref: "#/definitions/ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "405":
          description: Method Not Allowed
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "502":
          description: Bad Gateway
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Service Unavailable
          schema:
            $ref: "#/definitions/ErrorResponse"
        "504":
          description: Gateway Timeout
          schema:
            $ref: "#/definitions/ErrorResponse"
        "408":
          description: Request Timedout
          schema:
            $ref: "#/definitions/ErrorResponse"

     # Active group-policies will be retrieved 
...
definitions:
...
  NewParkingPolicy: 
    type: object
    required:
      - name
      - policyAuthorizerid
    properties:
      name:
        type: string
      policyAuthorizerid:
        type: string 
      description:
        type: string 
      timeZone:
        type: string
      tags:
        type: array
        items:
            type: string
      policyRule:
        type: array
        items:
         $ref: "#/definitions/PolicyRule"
...
  GetParkingPolicyResponse:
    type: object
    properties:
      uid:
        type: string
      policyAuthorizerid:
        type: string
      siteid:
        type: string
      orgid:
        type: string
      tags:
        type: array
        items:
            type: string
      isLibraryPolicy:
        type: boolean
      hashValue:
        type: string
      createdOn:
        type: number
      lastUpdated:
        type: number
      isDeleted:
        type: boolean
      version:
        type: number
      name:
        type: string
      description:
        type: string
      state:
        type: string
      timeZone:
        type: string
      lastupdated:
        type: number
      policyRule:
        type: array
        items:
         $ref: "#/definitions/PolicyRule"
  PolicyRule:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      priority:
        type: number
      parkingSchedule:
        $ref: "#/definitions/ParkingSchedule"
      parkingRule:
         $ref: "#/definitions/ParkingRule"
  ParkingSchedule:
    type: object
    properties:
      description:
        type: string
      daysOfWeek:
        type: array
        items:
           type: string
      months:
        type: array
        items:
           type: string
      datePeriod:
        $ref: "#/definitions/DatePeriod"
      timeRange:
        type: array
        items:
          $ref: "#/definitions/TimeRange"
...
  ParkingRule:
    type: object
    properties:
      description:
        type: string
      parkingAllowed:
        type: boolean
      parkingCharge:
        $ref: "#/definitions/ParkingCharge"
      parkingPenalty:
        $ref: "#/definitions/ParkingPenalty"
...
  CreateParkingPolicyResponse:
    $ref: "#/definitions/GetParkingPolicyResponse" 
  UpdateParkingPolicyResponse:
    $ref: "#/definitions/GetParkingPolicyResponse" 
  getAllParkingPolicyResponse:
    $ref: "#/definitions/GetParkingPolicyResponse" 
  DeleteParkingPolicyResponse:
    type: object
    properties:
      success:
        type: boolean
...
  GetParkingPolicyResponseByVersion:
    $ref: "#/definitions/GetParkingPolicyResponse" 
  GetParkingPolicyResponseByVersionHistory:
    type: object
    properties:
      policyStartDateTime:
        type: string
      policyEndDateTime:
        type: string
      createdOn:
        type: number
        format: int64
      version:
        type: number
        format: int32
  GetParkingPolicyResponseByTimeline:
    $ref: "#/definitions/GetParkingPolicyResponse"
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
  #   #required:
  #   #  - message
    #  properties:
    #    message:
    #      type: string